Add files via upload
 5-floor office building. Each floor is equipped with a 
different number of computers, like floor 1 has 11045, floor 2 has 54321, 
floor 3 has 234, floor 4 has 568, and floor 5 has 10. Configure the DHCP 
server on floor 1, the Email server should be connected on floor 3, the HTTP 
server should be connected on floor 2, and the DNS and FTP servers of the 
company are on floor 5. The organization requires a well-structured network to 
ensure efficient communication and scalability.